units tests passing WIP
